New and improved documentation
The CSS is from the Haiku User Guide, the general layout as I've
used them for other docs before.

Not much could be copied from the original ReadMe.html, other
than the section about the "Add file extension" under the
screenshot of the "Renaming Modes".

That screenshot uses JavaScript to change the image with a click.
Hope that's OK for everyone. There a no-script alternative.